refactor: 불필요한 wrapper 제거
