--- src/main/kotlin/lotto/domain/LottoProfitResult.kt
@@ -1,9 +1,0 @@
-package lotto.domain
-
-data class LottoProfitResult(
-    val totalTicketPrice: Int,
-    val totalPrize: Long,
-    val remainder: Int,
-) {
-    val profitRate: Double get() = String.format("%.2f", totalPrize.toDouble() / totalTicketPrice.toDouble()).toDouble()
-}
