+++ src/main/kotlin/lotto/domain/LottoResult.kt
@@ -1,7 +1,9 @@
 package lotto.domain
 
 data class LottoResult(
-    val lottoProfitResult: LottoProfitResult,
+    val totalTicketPrice: Int,
+    val totalPrize: Long,
+    val remainder: Int,
 ) {
-    val profitRate: Double get() = lottoProfitResult.profitRate
+    val profitRate: Double get() = String.format("%.2f", totalPrize.toDouble() / totalTicketPrice.toDouble()).toDouble()
 }
