+++ src/main/kotlin/lotto/view/ResultView.kt
@@ -2,9 +2,9 @@ package lotto.view
 
 import lotto.domain.LottoResult
 
-class ResultView() {
+class ResultView {
     fun printResult(lottoResult: LottoResult) {
-        println("총 수익은 ${lottoResult.lottoProfitResult.totalPrize} 입니다.")
+        println("총 수익은 ${lottoResult.totalPrize} 입니다.")
         println("총 수익률은 ${lottoResult.profitRate} 입니다.")
     }
 }
