--- src/test/kotlin/lotto/domain/LottoProfitResultTest.kt
@@ -1,16 +1,0 @@
-package lotto.domain
-
-import io.kotest.core.spec.style.StringSpec
-import io.kotest.matchers.shouldBe
-
-class LottoProfitResultTest : StringSpec({
-    "lotto profit result should calculate correct profit rate" {
-        val lottoProfitResult = LottoProfitResult(
-            totalTicketPrice = 3_000,
-            totalPrize = 5_000,
-            remainder = 0
-        )
-
-        lottoProfitResult.profitRate shouldBe 1.67
-    }
-})
