+++ src/test/kotlin/lotto/domain/LottoResultTest.kt
@@ -6,11 +6,15 @@ import io.kotest.matchers.shouldBe
 class LottoResultTest : StringSpec({
     "LottoResult should provide correct profit rate" {
         LottoResult(
-            LottoProfitResult(
-                totalTicketPrice = 5_000,
-                totalPrize = 4_600_000,
-                remainder = 0
-            )
+            totalTicketPrice = 5_000,
+            totalPrize = 4_600_000,
+            remainder = 0
         ).profitRate shouldBe 920
+
+        LottoResult(
+            totalTicketPrice = 3_000,
+            totalPrize = 5_000,
+            remainder = 0
+        ).profitRate shouldBe 1.67
     }
 })
